+++ page-templates/page-about-us-organization.php
@@ -40,7 +40,7 @@ if ( have_posts() ) :
 		get_template_part( 'template-parts/modules/hero-section' );
 		get_template_part( 'template-parts/modules/intro-cta', null, array( 'prefix' => 'organization_why_' ) );
 		get_template_part( 'template-parts/modules/button-text', null, array( 'prefix' => 'organigramm_' ) );
-		get_template_part( 'template-parts/pages/organization/team' );
+		get_template_part( 'template-parts/pages/about-us-organization/team' );
 		get_template_part( 'template-parts/modules/button-text', null, array( 'prefix' => 'organization_transparency_' ) );
 		get_template_part( 'template-parts/modules/cta-form', null, array( 'prefix' => 'organization_' ) );